# Better strategy for attention matrix multiplications when generating tokens 